AI/ai-django-mask/ai_django_mask/requirements.txt
absl-py==0.9.0
aioredis==1.3.1
appdirs==1.4.3
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1592528583041/work
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1598374659300/work
audioread==2.1.8
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1598352799320/work
channel==0.1.0
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography @ file:///C:/ci/cryptography_1598873979408/work
cycler==0.10.0
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1598956214003/work
django-cors-headers==3.5.0
django-extensions==3.0.9
django-redis==4.12.1
djangorestframework==3.11.1
entrypoints==0.3
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1599595709135/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1597406403308/work
h5py==2.10.0
hiredis==1.1.0
hyperlink==20.0.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
imutils @ file:///D:/bld/imutils_1594126159282/work
incremental==17.5.0
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598883894321/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi @ file:///C:/ci/jedi_1596472767194/work
Jinja2==2.11.2
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1593624380152/work
jsonschema @ file:///C:/ci/jsonschema_1594363671836/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa @ file:///home/conda/feedstock_root/build_artifacts/librosa_1595413403988/work
llvmlite==0.34.0
Markdown @ file:///C:/ci/markdown_1597415185889/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///C:/ci/matplotlib-base_1597858387645/work
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.1.0
mkl-random==1.1.0
mkl-service==2.3.0
msgpack==1.0.0
nbconvert @ file:///C:/ci/nbconvert_1594372737468/work
nbformat==5.0.7
notebook @ file:///C:/ci/notebook_1596837179121/work
numba @ file:///D:/bld/numba_1599085049111/work
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.4
pandas @ file:///C:/ci/pandas_1598371525254/work
pandocfilters==1.4.2
parso==0.7.0
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow @ file:///C:/ci/pillow_1594298234712/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1599743645983/work
prometheus-client==0.8.0
prompt-toolkit==1.0.14
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAudio==0.2.11
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
PyHamcrest==2.0.2
PyInquirer==1.0.3
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.16.0
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
python-magic-bin==0.4.14
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole @ file:///tmp/build/80754af9/qtconsole_1598374667791/work
QtPy==1.9.0
redis==3.5.3
regex==2020.7.14
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
resampy==0.2.2
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
scikit-learn @ file:///D:/bld/scikit-learn_1596546337481/work
scipy==1.4.1
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
SoundFile==0.10.2
sqlparse==0.3.1
tensorboard==2.0.2
tensorboard-plugin-wit==1.6.0
tensorflow==2.0.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.0.1
tensorflow-hub==0.7.0
tensorflowjs==2.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.2
